refactor: Modularizar arquitectura de la aplicación
Fase A: Los hooks en src/hooks/supabase/ ya estaban divididos
- Eliminado useSupabase.jsx original (2,095 líneas)
- 13 hooks individuales funcionando correctamente

Fase B: Refactorizado App.jsx (1,251 → 251 líneas)
- Creado useAppState.js: Estados consolidados de la app
- Creado useAppHandlers.js: Handlers extraídos
- Creado AppModals.jsx: Renderizado de modales

Fase C: Refactorizado VistaPedidos.jsx (871 → 167 líneas)
- Extraído PedidoCard.jsx: Card individual de pedido
- Extraído PedidoFilters.jsx: Filtros de búsqueda
- Extraído PedidoActions.jsx: Dropdown de acciones (memoizado)
- Extraído PedidoStats.jsx: Resumen de estadísticas
- Extraído VistaRutaTransportista.jsx: Vista móvil para transportistas
