+++ src/components/pedidos/index.js
@@ -1,0 +1,8 @@
+/**
+ * Exportaciones de componentes de pedidos
+ */
+export { default as PedidoCard } from './PedidoCard';
+export { default as PedidoActions } from './PedidoActions';
+export { default as PedidoFilters } from './PedidoFilters';
+export { default as PedidoStats } from './PedidoStats';
+export { default as VistaRutaTransportista } from './VistaRutaTransportista';
